requirements.txt
langchain>=0.2.0
langchain-text-splitters>=0.2.0
sentence-transformers>=3.0.0
pinecone>=5.4.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.36.0
pypdf>=4.2.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
openai>=1.40.0
requests>=2.32.0
tqdm>=4.66.0
pytest>=8.3.0
